Auto-detect apt caching proxies when building the images
We are using the latest (beta) version of the auto-apt-proxy script.
(This script is installable via apt, but we are using a newer
bleeding-edge version here, since it supports finding the default
gateway using busybox.)

This is the version from the [parallel-detection][1] branch.
The precise commit (made 2020-12-16) it came from is:
git+https://salsa.debian.org/debian/auto-apt-proxy.git#d5b235b7

[1]: <https://salsa.debian.org/debian/auto-apt-proxy/-/tree/parallel-detection>